requirements.txt
bcrypt==3.2.0
Django==3.0.6
django-rest-framework==0.1.0
djangorestframework==3.11.0
PyJWT==1.7.1
python-dotenv==0.14.0
psycopg2==2.8.6
uwsgi==2.0.19.1
django-cors-headers==3.5.0
google-api-python-client==1.12.4
django-storages==1.10.1
boto3==1.16.7
python-magic-bin==0.4.14
celery==4.7.7
django-queued-storage==0.8.0
